feat: the alert was created

+++ Frontend/src/views/Tasks.vue
@@ -286,6 +286,7 @@ import {
   where,
 } from "firebase/firestore";
 import { useRoute } from "vue-router";
+import { alertQuestion, alertSuccess } from "@/components/alert";
 
 // Estado de la aplicación
 const showAddTaskModal = ref(false);
@@ -405,11 +406,15 @@ const saveTask = async () => {
     progress: 0,
     status: "todo",
   });
+  alertSuccess("tarea guardada")
 };
 
 // Eliminar tarea
 const deleteTask = async (id) => {
+  const result= await alertQuestion("¿deseas eliminar la tarea?")
+  if(!result.isConfirmed) return;
   await deleteDoc(doc(db, "tasks", id));
+  alertSuccess("tarea eliminada")
 };
 
 // Editar tarea
@@ -451,6 +456,8 @@ const editNote = (note) => {
 };
 
 const deleteNote = async (noteId) => {
+  const result= await alertQuestion("¿quieres eliminar la nota?")
+  if(!result.isConfirmed) return;
   await deleteDoc(doc(db, "notes", noteId));
 };
 
@@ -471,6 +478,7 @@ const saveNote = async () => {
       date: new Date().toLocaleDateString(),
       userId,
     });
+    alertSuccess("nota creada")
   }
   resetNoteForm();
   showAddNoteModal.value = false;
